fix(wallet): update savings/spending wallet screen UI
