--- Bitkit/Views/Wallets/SpendingWalletView.swift
@@ -1,128 +1,0 @@
-import SwiftUI
-
-struct SpendingWalletView: View {
-    @EnvironmentObject var activity: ActivityListViewModel
-    @EnvironmentObject var app: AppViewModel
-    @EnvironmentObject var navigation: NavigationViewModel
-    @EnvironmentObject var wallet: WalletViewModel
-
-    /// Whether there are any lightning activities to display
-    private var hasLightningActivities: Bool {
-        guard let activities = activity.lightningActivities else { return false }
-        return !activities.isEmpty
-    }
-
-    /// tab bar + spacing
-    private var bottomPadding: CGFloat {
-        64 + 32
-    }
-
-    var body: some View {
-        ZStack(alignment: .top) {
-            VStack(spacing: 0) {
-                NavigationBar(title: t("wallet__spending__title"), icon: "ln")
-
-                MoneyStack(
-                    sats: wallet.totalLightningSats,
-                    showSymbol: true,
-                    showEyeIcon: false,
-                    enableSwipeGesture: true,
-                    enableHide: true,
-                    testIdPrefix: "TotalBalance"
-                )
-                .padding(.top)
-
-                if wallet.balanceInTransferToSpending > 0 {
-                    IncomingTransfer(amount: UInt64(wallet.balanceInTransferToSpending))
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .padding(.top, 16)
-                }
-
-                if wallet.totalLightningSats > 0 || hasLightningActivities {
-                    if wallet.totalLightningSats > 0, let channels = wallet.channels, !channels.isEmpty {
-                        transferButton
-                            .transition(.move(edge: .leading).combined(with: .opacity))
-                            .padding(.top, 32)
-                    }
-
-                    ScrollView(showsIndicators: false) {
-                        ActivityList(viewType: .lightning)
-
-                        CustomButton(title: t("common__show_all"), variant: .tertiary) {
-                            navigation.navigate(.activityList)
-                        }
-                        .padding(.bottom, bottomPadding)
-                    }
-                    .accessibilityIdentifier("HomeScrollView")
-                    .refreshable {
-                        do {
-                            try await wallet.sync()
-                            try await activity.syncLdkNodePayments()
-                        } catch {
-                            app.toast(error)
-                        }
-                    }
-                    .frame(maxWidth: .infinity, minHeight: 400)
-                    .transition(.move(edge: .leading).combined(with: .opacity))
-                } else {
-                    Spacer()
-                    WalletOnboardingView(type: .spending)
-                        .padding(.bottom, bottomPadding)
-                        .transition(.move(edge: .trailing).combined(with: .opacity))
-                }
-            }
-            .padding(.horizontal)
-            .background(alignment: .topTrailing) {
-                Image("coin-stack-x-2")
-                    .resizable()
-                    .frame(width: 256, height: 256)
-                    .offset(x: 128)
-            }
-
-            // Bottom gradient: black 0% to black 100%
-            VStack {
-                Spacer()
-                LinearGradient(
-                    colors: [.black.opacity(0), .black],
-                    startPoint: .top,
-                    endPoint: .bottom
-                )
-                .frame(height: 140)
-            }
-            .ignoresSafeArea()
-            .allowsHitTesting(false)
-        }
-        .navigationBarHidden(true)
-        .animation(.spring(response: 0.3), value: wallet.totalLightningSats)
-    }
-
-    var transferButton: some View {
-        CustomButton(
-            title: t("lightning__transfer_to_savings_button"),
-            variant: .secondary,
-            icon: Image("arrow-up-down")
-                .resizable()
-                .scaledToFit()
-                .frame(width: 16, height: 16)
-                .foregroundColor(.white80)
-        ) {
-            if app.hasSeenTransferToSavingsIntro {
-                navigation.navigate(.savingsAvailability)
-            } else {
-                navigation.navigate(.savingsIntro)
-            }
-        }
-        .accessibilityIdentifier("TransferToSavings")
-    }
-}
-
-#Preview {
-    NavigationStack {
-        SpendingWalletView()
-    }
-    .environmentObject(WalletViewModel())
-    .environmentObject(AppViewModel())
-    .environmentObject(CurrencyViewModel())
-    .environmentObject(ActivityListViewModel())
-    .preferredColorScheme(.dark)
-}
